+++ Tests/NukePerformanceTests/ImagePipelinePerformanceTests.swift
@@ -112,10 +112,6 @@ private func makePipeline(_ configure: (inout ImagePipeline.Configuration) -> Vo
 
         $0.isDecompressionEnabled = false
 
-        // This must be off for this test, because rate limiter is optimized for
-        // the actual loading in the apps and not the synthetic tests like this.
-        $0.isRateLimiterEnabled = false
-
         // Remove decoding from the equation
         $0.makeImageDecoder = { _ in ImageDecoders.Empty() }